added layer info

+++ filetree/tree.go
@@ -208,9 +208,10 @@ func (tree *FileTree) MarkRemoved(path string) error {
 	return node.AssignDiffType(Removed)
 }
 
-func StackRange(trees []*FileTree, index uint) *FileTree {
+func StackRange(trees []*FileTree, index int) *FileTree {
+	// TMP TMP TMP: TODO: later change the index of both of these to 0
 	tree := trees[1].Copy()
-	for idx := uint(1); idx <= index; idx++ {
+	for idx := 1; idx <= index; idx++ {
 		tree.Stack(trees[idx])
 	}
 	return tree
